Add PHONE_VERSION endpoint handler
Add a phone version endpoint handler so that we respond appropriately when
connecting to a pebble.  The pebble sends this query at the start of an RFCOMM
connection and will assume a bare minimum of supported features if we don't
respond.

The only feature this controls (that I can tell) is whether or not the music
app sends next/previous/pause commands back to us.  Soo...this fixes the music
app so that it actually works.  Yay!